Package rename: blib -> speclib
