+++ src/org/labkey/targetedms/TargetedMSController.java
@@ -182,9 +182,9 @@ import org.labkey.targetedms.parser.SampleFileChromInfo;
 import org.labkey.targetedms.parser.SkylineBinaryParser;
 import org.labkey.targetedms.parser.SkylineDocumentParser;
 import org.labkey.targetedms.parser.TransitionChromInfo;
-import org.labkey.targetedms.parser.blib.BlibSpectrumReader;
 import org.labkey.targetedms.parser.list.ListDefinition;
 import org.labkey.targetedms.parser.skyaudit.AuditLogEntry;
+import org.labkey.targetedms.parser.speclib.BlibSpectrumReader;
 import org.labkey.targetedms.pipeline.ChromatogramCrawlerJob;
 import org.labkey.targetedms.query.ChromatogramDisplayColumnFactory;
 import org.labkey.targetedms.query.ConflictResultsManager;
