+++ src/org/labkey/targetedms/view/spectrum/LibrarySpectrumMatchGetter.java
@@ -30,8 +30,8 @@ import org.labkey.targetedms.parser.Peptide;
 import org.labkey.targetedms.parser.PeptideSettings;
 import org.labkey.targetedms.parser.Precursor;
 import org.labkey.targetedms.parser.SampleFile;
-import org.labkey.targetedms.parser.blib.BlibSpectrum;
-import org.labkey.targetedms.parser.blib.BlibSpectrumReader;
+import org.labkey.targetedms.parser.speclib.BlibSpectrum;
+import org.labkey.targetedms.parser.speclib.BlibSpectrumReader;
 import org.labkey.targetedms.query.IsotopeLabelManager;
 import org.labkey.targetedms.query.LibraryManager;
 import org.labkey.targetedms.query.ModificationManager;
